+++ test/dummy/app/models/workspace.rb
@@ -1,3 +1,3 @@
 class Workspace < ApplicationRecord
-  recording_studio_recordable label: "Workspace", root: true
+  recording_studio_recordable label: "Workspace", root: true, allowed_parent_types: []
 end