+++ test/dummy/test/integration/recording_studio_v3_host_test.rb
@@ -3,10 +3,11 @@
 require "test_helper"
 
 class RecordingStudioV3HostTest < ActiveSupport::TestCase
-  test "dummy app loads root switchable config and controller support" do
+  test "dummy app loads root switchable config and default layout" do
     assert_equal [ "all_workspaces" ], RecordingStudioRootSwitchable.configuration.scopes.keys
     assert_equal :application_layout, RecordingStudioRootSwitchable.configuration.layout
     assert_includes ApplicationController.ancestors, RecordingStudio::RootSwitchable::ControllerSupport
+    assert_includes ApplicationController.ancestors, RecordingStudio::UsesDefaultLayout
   end
 
   test "dummy app validates v3 recordable declarations" do